requirements.txt
jxmlease==1.0.3
requests==2.28.2

PyYAML~=6.0
pyxtension~=1.14.5
DateTime~=5.0